Refractor for clarity and readablility
