+++ downloader.py
@@ -15,11 +15,8 @@ import sys
 import getopt
 import logging
 import traceback
-import pandas as pd
-import mimetypes
 
-from duploader.dupload import Dupload
-from duploader.utils import Utils
+from wikiutils.utils import Utils
 
 if __name__ == "__main__":
     format = "%(asctime)s: %(message)s"
@@ -27,14 +24,13 @@ if __name__ == "__main__":
 
 
 # Helper classes
-duploader = Dupload()
 utils = Utils()
 
 # Input parameters
 download_limit = 0  # Total number of bytes to download
 batch_size = 0  # Size of download batches (in bytes)
 input_df = ""  # input of parquet files generated by ingestion3 wiki job
-save_location = ""  # output location
+base_output_path = ""  # output location
 
 # Controlling vars
 df_rows = list()  #
@@ -60,22 +56,27 @@ for opt, arg in opts:
     elif opt in ("-i", "--input"):
         input_df = arg
     elif opt in ("-o", "--output"):
-        save_location = arg.rstrip('/')
+        base_output_path = arg.rstrip('/')
     elif opt in ("-b", "--batch_size"):
         batch_size = int(arg)
 
 logging.info(f"Size limit: {utils.sizeof_fmt(download_limit)}")  # 1 TB === 1000000000000, 0 for no limit
 logging.info(f"Batch size: {utils.sizeof_fmt(batch_size)}")  # 1 TB === 1000000000000
 logging.info(f"Input:      {input_df}")
-logging.info(f"Output:     {save_location}")
+logging.info(f"Output:     {base_output_path}")
 
+# Get individual parquet files from ingestion3 wiki output
 file_list = utils.get_parquet_files(path=input_df)
-df_output_path = f"{save_location}/batch_{batch_number}/data/"
-record_count = 1
+
+df_batch_out = f"{base_output_path}/batch_{batch_number}/data/"
+dpla_item_count = 1
 
 for parquet_file in file_list:
     logging.info(f"Processing...{parquet_file}")
+    utils.create_path(df_batch_out)
+    # read parquet file
     df = utils.get_df(parquet_file, columns=columns)
+
     for row in df.itertuples(index=['id', 'wiki_markup', 'iiif', 'media_master', 'title']):
         try:
             dpla_id = getattr(row, 'id')
@@ -87,7 +88,7 @@ for parquet_file in file_list:
             logging.error(f"Unable to get attributes from row {row}: {e}. Aborting...")
             break
 
-        # Are we working with iiif or media_master?
+        # Are we working with IIIF or media_master?
         download_urls = utils.get_iiif_urls(iiif) if iiif else media_master
 
         """
@@ -103,11 +104,8 @@ for parquet_file in file_list:
             - size of asset 
             - Wiki markup 
         """
-
-
-
         # TODO `out` should be the root save location or the asset path?
-        out, time, size = save_location, 0, 0  # Defaults
+        out, time, size = base_output_path, 0, 0  # Defaults
         asset_count = 1
 
         # MULTI-ASSET SUPPORT
@@ -115,17 +113,14 @@ for parquet_file in file_list:
         rows = list()
         for url in download_urls:
             # Create asset path
-            asset_path = f"{save_location}/batch_{batch_number}/assets/{dpla_id[0]}/{dpla_id[1]}/{dpla_id[2]}/{dpla_id[3]}/{dpla_id}/{asset_count}_{dpla_id}"
+            asset_path = f"{base_output_path}/batch_{batch_number}/assets/" \
+                f"{dpla_id[0]}/{dpla_id[1]}/{dpla_id[2]}/{dpla_id[3]}/{dpla_id}/{asset_count}_{dpla_id} "
 
             utils.create_path(asset_path)
-            utils.create_path(df_output_path)
-
-            # The asset path should be {0}/{1}/{2}/{3}/{dpla_id}/{assetcount_dplaId}
-            #
             try:
-                out, time, size = duploader.download_single_item(url=url, save_location=asset_path)
-                # name is ignored
+                out, time, size = utils.download(url=url, out=asset_path)
             except Exception as e:
+                # If a single asset files for a multi-asset upload then all assets are dropped
                 logging.error(f"Aborting all assets for {dpla_id}\n{e}\n{traceback.format_exc()}")
                 rows = list()
                 out = None
@@ -155,18 +150,18 @@ for parquet_file in file_list:
                     'page': asset_count
                 }
                 rows.append(row)
-            logging.info(f"Page {asset_count} for {dpla_id}")
+
             asset_count = asset_count + 1  # increment asset count
 
         # append all assets/rows for a given metadata record
-        logging.info(f"Appending {len(rows)} rows to df")
-        df_rows.extend(rows)
+        if len(rows) > 0:
+            df_rows.extend(rows)
 
         # Only log a message every 100 records
-        if(record_count % 100 == 0):
-            logging.info(f"{record_count} records and {len(df_rows)} images")
+        if dpla_item_count % 100 == 0:
+            logging.info(f"{dpla_item_count} records and {len(df_rows)} assets")
 
-        record_count = record_count + 1
+        dpla_item_count = dpla_item_count + 1
 
         if batch_downloaded > batch_size:
             logging.info(f"Upload quota met for batch {batch_number}")
@@ -175,14 +170,14 @@ for parquet_file in file_list:
                              f"\t{utils.sizeof_fmt(batch_downloaded)}")
 
             # Save upload info dataframe
-            batch_parquet_out_path = f"{df_output_path}batch_{batch_number}.parquet"
+            batch_parquet_out_path = f"{df_batch_out}batch_{batch_number}.parquet"
             utils.write_parquet(batch_parquet_out_path, df_rows, upload_parquet_columns)
 
             # Reset batch control vars, update df_output_path
             df_rows = list()
             batch_number = batch_number + 1
             batch_downloaded = 0
-            df_output_path = f"{save_location}/batch_{batch_number}/data/"
+            df_batch_out = f"{base_output_path}/batch_{batch_number}/data/"
             logging.info(f"Starting batch number {batch_number}")
 
         # If there is a total limit in place then abort after it has been breached.
@@ -192,7 +187,7 @@ for parquet_file in file_list:
                              f"\t{len(df_rows)} files \n" \
                              f"\t{utils.sizeof_fmt(batch_downloaded)}")
 
-            batch_parquet_out_path = f"{df_output_path}batch_{batch_number}.parquet"
+            batch_parquet_out_path = f"{df_batch_out}batch_{batch_number}.parquet"
             utils.write_parquet(batch_parquet_out_path, df_rows, upload_parquet_columns)
             df_rows = list()
             logging.info("Exiting...")
@@ -200,6 +195,6 @@ for parquet_file in file_list:
 
 # If finished processing parquet files without breaching limits then write data out
 if df_rows:
-    batch_parquet_out_path = f"{df_output_path}batch_{batch_number}.parquet"
+    batch_parquet_out_path = f"{df_batch_out}batch_{batch_number}.parquet"
     utils.write_parquet(batch_parquet_out_path, df_rows, upload_parquet_columns)
     df_rows = list()  # reset
